feat: add Consume.create_stage for LazyStage
